+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 * Support allOf definition (#11)
 * Support wildcard status code (#12)
+* Support wildcard content type (#13)
 
 ## 0.1.5 (2018-12-23)
 * First release for production usage